+++ changelog.md
@@ -2,6 +2,8 @@
 ##### 0.2.1
 - **Changed:**
   - Adjusted dark mouse over colors to be more visible
+- **Fixed:**
+  - Fix issues with closing client and still being connected and saying no to close
 ##### 0.2.0 - 2017-05-19
 - **New:**
   - Add ${variable.FUNCTION} format support to allow similar format to javascript.